Fixed issue #67
