+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "kunta-api-www",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "www visualization layer for kunta-api data",
   "main": "app.js",
   "scripts": {
